add spring security

+++ andrei-s/src/main/java/org/java/mentorship/andrei_s/controller/AlbumController.java
@@ -1,9 +1,11 @@
 package org.java.mentorship.andrei_s.controller;
 
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.java.mentorship.andrei_s.domain.Album;
 import org.java.mentorship.andrei_s.service.AlbumService;
 import org.java.mentorship.andrei_s.service.SongService;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -33,19 +35,22 @@ public class AlbumController {
     }
 
     @PostMapping()
-    public ResponseEntity<Album> createNew(@RequestBody Album album) {
+    public ResponseEntity<Album> createNew(HttpServletRequest request, @RequestBody Album album) {
+        if (!request.isUserInRole("ADMIN")) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
         Album newAlbum = albumService.createNew(album);
         return ResponseEntity.ok(newAlbum);
     }
 
     @DeleteMapping("/{id}")
-    ResponseEntity<String> delete(@PathVariable(name = "id") int id) {
+    ResponseEntity<String> delete(HttpServletRequest request, @PathVariable(name = "id") int id) {
+        if (!request.isUserInRole("ADMIN")) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
         albumService.deleteById(id);
         return ResponseEntity.ok("Deleted album");
     }
 
     @PutMapping("/{id}")
-    ResponseEntity<Album> modify(@PathVariable(name = "id") int id, @RequestBody Album album) {
+    ResponseEntity<Album> modify(HttpServletRequest request, @PathVariable(name = "id") int id, @RequestBody Album album) {
+        if (!request.isUserInRole("ADMIN")) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
         return ResponseEntity.ok(albumService.updateById(id, album));
     }
 }
