+++ andrei-s/src/main/java/org/java/mentorship/andrei_s/controller/SongController.java
@@ -1,8 +1,10 @@
 package org.java.mentorship.andrei_s.controller;
 
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.java.mentorship.andrei_s.domain.Song;
 import org.java.mentorship.andrei_s.service.SongService;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -36,19 +38,22 @@ public class SongController {
     }
 
     @PostMapping()
-    public ResponseEntity<Song> createNew(@RequestBody Song song) {
+    public ResponseEntity<Song> createNew(HttpServletRequest request, @RequestBody Song song) {
+        if (!request.isUserInRole("ADMIN")) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
         Song newSong = songService.createNew(song);
         return ResponseEntity.ok(newSong);
     }
 
     @DeleteMapping("/{id}")
-    ResponseEntity<String> delete(@PathVariable(name = "id") int id) {
+    ResponseEntity<String> delete(HttpServletRequest request, @PathVariable(name = "id") int id) {
+        if (!request.isUserInRole("ADMIN")) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
         songService.deleteById(id);
         return ResponseEntity.ok("Deleted song");
     }
 
     @PutMapping("/{id}")
-    ResponseEntity<Song> modify(@PathVariable(name = "id") int id, @RequestBody Song song) {
+    ResponseEntity<Song> modify(HttpServletRequest request, @PathVariable(name = "id") int id, @RequestBody Song song) {
+        if (!request.isUserInRole("ADMIN")) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
         return ResponseEntity.ok(songService.updateById(id, song));
     }
